[Build] perform create-sources and attach-javaDoc in 'verify'-phase
Execute the javadoc:jar and source:jar goals in the the verify phase
instead of the deploy phase. If java-doc and sources should not be
attached then simply execute the default life-cycle only up to a phase
before verify (like mvn test or package).

Because now the jar goals are executed more frequently they should run
quietly and only show warnings/errors.

Furthermore since the mentioned goals are not bound to the deploy phase
anymore there is no need to disable the execution of the
maven-deploy-plugin. If deployment is not intended then then simply do
not run Maven up to the deploy phase.